requirements.txt
Django>=5.1
djangorestframework>=3.15.0
django-mongodb-backend==5.1.0b1
drf-yasg==1.21.10